Fixes servo speed management
Corrects servo speed handling by allowing null values to default to the position's defined speed.

Also, changes the type of FULL_SPEED and HALF_SPEED constants to byte for consistency.

+++ eurobot-system-lib-parent/eurobot-system-lib-core/src/main/java/org/arig/robot/services/AbstractCommonRobotServosService.java
@@ -40,8 +40,8 @@ public abstract class AbstractCommonRobotServosService extends AbstractServosSer
     protected static final String POS_ETAGE_2 = "Etage 2";
     protected static final String POS_REPOS = "Repos";
 
-    protected static final int FULL_SPEED = 0;
-    protected static final int HALF_SPEED = 30;
+    protected static final byte FULL_SPEED = 0;
+    protected static final byte HALF_SPEED = 30;
 
     protected static final String GROUP_INDIVIDUAL_AVANT = "Individuel avant";
     protected static final String GROUP_INDIVIDUAL_ARRIERE = "Individuel arrière";
